Replace seek_until_consume_chars with seek_until_chars
In some places it's desireable to seek and consume the character you
stop on, whereas in others it's not. In my mind, it's a bit cleaner to
default to not consuming that stop character and in the other cases,
explicitly consume it to eliminate confusion about what happening to the
stream (It confused me several times already).